requirements.txt
flask==2.0.1
flask-cors==3.0.10
Pillow==9.0.0
numpy==1.23.5
scikit-image==0.19.3
python-dotenv==0.19.0
flask-jwt-extended==4.3.1
werkzeug==2.0.3